% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connections.R
\name{get_data_mart_conn}
\alias{get_data_mart_conn}
\title{Returns a connection object for the data mart on the shiny server}
\usage{
get_data_mart_conn()
}
\description{
Returns a connection object for the data mart on the shiny server
}
