% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connections.R
\name{get_mpc_conn}
\alias{get_mpc_conn}
\title{Returns a connection object for mpc centricity}
\usage{
get_mpc_conn(driver = "ODBC Driver 17 for SQL Server")
}
\description{
Returns a connection object for mpc centricity
}
