% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connections.R
\name{get_mysql_conn}
\alias{get_mysql_conn}
\title{Get a MYSQL Connection}
\usage{
get_mysql_conn(database, ...)
}
\arguments{
\item{database}{Name of the database you are trying to connect to.}

\item{...}{arguments to pass into \code{DBI::dbConnect} function.}
}
\description{
Get a MYSQL Connection
}
