% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connections.R
\name{get_postgres_conn}
\alias{get_postgres_conn}
\title{Get a Postgres Connection}
\usage{
get_postgres_conn(database, ...)
}
\arguments{
\item{database}{Name of the database you are trying to connect to.}

\item{...}{arguments to pass into \code{DBI::dbConnect} function.}
}
\description{
Get a Postgres Connection
}
