% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connections.R
\name{get_smpc_conn}
\alias{get_smpc_conn}
\title{Returns a connection object for smpc centricity}
\usage{
get_smpc_conn(driver = "ODBC Driver 17 for SQL Server")
}
\description{
Returns a connection object for smpc centricity
}
