% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connections.R
\name{get_sql_server_conn}
\alias{get_sql_server_conn}
\title{Get a SQL Server Connection}
\usage{
get_sql_server_conn(database, ...)
}
\arguments{
\item{database}{Name of the database you are trying to connect to.}

\item{...}{arguments to pass into \code{DBI::dbConnect} function.}
}
\description{
Get a SQL Server Connection
}
